Extract EncryptionKeyPair to separate file
